# docs/ncc-api-swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "NCC - Client API",
    "description": "The NEM Community Client (NCC) provides an RESTful API to the wallet functionality. Thus enabling alternative front-end implementations be provided by the community.",
    "version": "0.6.37-BETA"
  },
  "host": "localhost:8989",
  "schemes": [
    "http"
  ],
  "basePath": "/ncc/api",
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/account/find": {
      "post": {
        "description": "Gets information about the specified account. Looks-up an account which is known by its address id",
        "tags": [
          "Account"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The address of the account to be found.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccountId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Details about the account.",
            "schema": {
              "$ref": "#/definitions/AccountViewModel"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/account/transactions/unconfirmed": {
      "post": {
        "description": "Gets information about the specified account and unconfirmed transactions.",
        "tags": [
          "Account",
          "Transaction"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The account identifier.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccountHashRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account and transactions information.",
            "schema": {
              "$ref": "#/definitions/AccountTransactionsPair"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/account/transactions/all": {
      "post": {
        "description": "Gets information about the specified account and incoming and outgoing confirmed and unconfirmed transactions.",
        "tags": [
          "Account",
          "Transaction"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The account (and optional) transaction identifier.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccountDatabaseIdRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account and transactions information.",
            "schema": {
              "$ref": "#/definitions/AccountTransactionsPair"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/account/transactions/confirmed": {
      "post": {
        "description": "Gets information about the specified account and incoming and outgoing confirmed transactions.",
        "tags": [
          "Account",
          "Transaction"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The account (and optional) transaction identifier.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccountDatabaseIdRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account and transactions information.",
            "schema": {
              "$ref": "#/definitions/AccountTransactionsPair"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/account/transactions/incoming": {
      "post": {
        "description": "Gets information about the specified account and incoming confirmed transactions.",
        "tags": [
          "Account",
          "Transaction"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The account (and optional) transaction identifier.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccountDatabaseIdRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account and transactions information.",
            "schema": {
              "$ref": "#/definitions/AccountTransactionsPair"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/account/transactions/outgoing": {
      "post": {
        "description": "Gets information about the specified account and outgoing confirmed transactions.",
        "tags": [
          "Account",
          "Transaction"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The account (and optional) transaction identifier.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccountDatabaseIdRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account and transactions information.",
            "schema": {
              "$ref": "#/definitions/AccountTransactionsPair"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/account/harvests": {
      "post": {
        "description": "Retrieves a list of infos on harvested blocks for an account.",
        "tags": [
          "Account",
          "Harvest"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The account (and optional) transaction identifier.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccountDatabaseIdRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account and transactions information.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/HarvestInfoViewModel"
              }
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/unlock": {
      "post": {
        "description": "Unlock the account on the connected NIS server (start harvesting).",
        "tags": [
          "Wallet",
          "Account",
          "Harvest"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The account / wallet view model",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccountWalletRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account is unlocked."
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/lock": {
      "post": {
        "description": "Lock the account on the connected NIS server (stop harvesting).",
        "tags": [
          "Wallet",
          "Account",
          "Harvest"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The account / wallet view model",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccountWalletRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account is locked."
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/remote/unlock": {
      "post": {
        "description": "Unlock the account on the remote NIS server (start delegated harvesting). Remote address being used has to be announced previously.",
        "tags": [
          "Wallet",
          "Account",
          "Harvest"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The remote harvester view model",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccountWalletPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account is unlocked on the remote server."
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/remote/lock": {
      "post": {
        "description": "Lock the account on the remote NIS server (stop delegated harvesting).",
        "tags": [
          "Wallet",
          "Account",
          "Harvest"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The remote harvester view model",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccountWalletRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account is locked on the remote server."
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/remote/status": {
      "post": {
        "description": "Checks if account is unlocked on remote server.",
        "tags": [
          "Wallet",
          "Account",
          "Information"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The remote harvester view model",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccountWalletRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account and transactions information.",
            "schema": {
              "$ref": "#/definitions/AccountStatusViewModel"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponseHttp": {
      "properties": {
        "timeStamp": {
          "type": "integer",
          "description": "Point in time when error happened. Time is based on NEM epoch time."
        },
        "error": {
          "type": "string",
          "description": "Error message."
        },
        "message": {
          "type": "string",
          "description": "Additional message about the error."
        },
        "status": {
          "$ref": "#/definitions/HttpErrorCodes",
          "description": "Status of the error."
        }
      },
      "example": {
        "timeStamp": 4978518,
        "error": "Not Found",
        "message": null,
        "status": 404
      }
    },
    "ErrorResponseNcc": {
      "properties": {
        "timeStamp": {
          "type": "integer",
          "description": "Point in time when error happened. Time is based on NEM epoch time."
        },
        "error": {
          "type": "string",
          "description": "Error message."
        },
        "message": {
          "type": "string",
          "description": "Additional message about the error."
        },
        "status": {
          "$ref": "#/definitions/NccErrorCodes",
          "description": "Status of the error."
        }
      },
      "example": {
        "timeStamp": 4978518,
        "error": "Not Found",
        "message": null,
        "status": 404
      }
    },
...
    "HttpErrorCodes": {
      "type": "integer",
      "format": "int32",
      "description": "Error codes that indicate the reason for failing:  \n* 400 - Bad request, e.g., mandatory parameter is missing or value is not correct.\n* 401 - Unauthorized, Request must be authorized.\n* 404 - Not Found, requested resource not found.",
      "enum": [
        400,
        401,
        404,
        500
      ]
    },
...
    "WalletNamePasswordPair": {
      "properties": {
        "wallet": {
          "type": "string",
          "description": "The name of the wallet."
        },
        "password": {
          "type": "string",
          "description": "The password used for encrypting the wallet."
        }
      },
      "description": "The name and password of a wallet."
    },
...
    "AccountViewModel": {
      "properties": {
        "address": {
          "type": "string",
          "description": "The address of the account."
        },
        "remoteStatus": {
          "type": "string",
          "description": "The remote status of the account.\n* REMOTE - The account is remote one, and therefore RemoteStatus is not applicable for it.\n* REMOTE - The account has activated remote harvesting (but not yet active).\n* ACTIVE - The account has activated remote harvesting and remote harvesting is active.\n* DEACTIVATING - The account has deactivated remote harvesting (but remote harvesting is still active).\n* INACTIVE - The account has inactive remote harvesting, or it has deactivated remote harvesting and deactivation is operational."
        },
        "publicKey": {
          "type": "string",
          "description": "The public key of the account."
        },
        "balance": {
          "type": "integer",
          "format": "int64",
          "description": "The amount of NEMs (in uNEM) of the account."
        },
        "vestedBalance": {
          "type": "integer",
          "format": "int64",
          "description": "The vested balance of the account (in uNEM: 1.000.000uNEM = 1 NEM)."
        },
        "importance": {
          "type": "number",
          "format": "double",
          "description": "The importance of the account to the NEM network."
        },
        "harvestedBlocks": {
          "type": "integer",
          "format": "int64",
          "description": "The number of harvested blocks."
        },
        "status": {
          "type": "string",
          "description": "The status of the account.\n*UNKNOWN - The account status is unknown (needed for ncc).\n*LOCKED - The account is locked.\n*UNLOCKED - The account is unlocked"
        },
        "multisigAccounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AccountInfo"
          },
          "description": "list of accounts (info) that this account is cosignatory of."
        },
        "cosignatories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AccountInfo"
          },
          "description": "List of accounts (info) that are cosignatories of this account."
        },
        "remotePublicKey": {
          "type": "string",
          "description": "The account's remote public key (optional)."
        },
        "remoteAddress": {
          "type": "string",
          "description": "The account's remote address (optional)."
        }
      },
      "description": "The content of the wallet."
    },
    "AccountInfo": {
      "properties": {
        "address": {
          "type": "string",
          "description": "The address of the account."
        },
        "publicKey": {
          "type": "string",
          "description": "The public key of the account."
        },
        "balance": {
          "type": "integer",
          "format": "int64",
          "description": "The amount of NEMs (in uNEM) of the account."
        },
        "vestedBalance": {
          "type": "integer",
          "format": "int64",
          "description": "The vested balance of the account (in uNEM: 1.000.000uNEM = 1 NEM)."
        },
        "importance": {
          "type": "number",
          "format": "double",
          "description": "The importance of the account to the NEM network."
        },
        "harvestedBlocks": {
          "type": "integer",
          "format": "int64",
          "description": "The number of harvested blocks."
        },
        "label": {
          "type": "string",
          "description": "The label of the account (public label)."
        }
      },
      "description": "list of accounts (info) that this account is cosignatory of."
    },
...
    "AccountId": {
      "description": "View model that represents an account id information.",
      "properties": {
        "account": {
          "type": "string"
        }
      }
    },
    "AccountHashRequest": {
      "description": "A request representing an account and an optional hash.",
      "properties": {
        "hash": {
          "type": "string"
        }
      }
    },
    "AccountTransactionsPair": {
      "description": "A request representing an account and an optional hash.",
      "allOf": [
        {
          "$ref": "#/definitions/AccountViewModel"
        },
        {
          "properties": {
            "transactions": {
              "description": "The transactions.",
              "type": "array",
              "items": {
                "$ref": "#/definitions/TransactionViewModel"
              }
            }
          }
        }
      ]
    },
...
    "AccountDatabaseIdRequest": {
      "description": "A request representing an account and an optional database id.",
      "allOf": [
        {
          "$ref": "#/definitions/AccountId"
        },
        {
          "properties": {
            "id": {
              "description": "The database id.",
              "type": "integer",
              "format": "int64"
            }
          }
        }
      ]
    },
    "HarvestInfoViewModel": {
      "description": "A view model representing harvest information.",
      "properties": {
        "id": {
          "description": "The block id.",
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string",
          "description": "A text stating the block height."
        },
        "timeStamp": {
          "description": "The time stamp.",
          "type": "integer",
          "format": "int64"
        },
        "fee": {
          "description": "The total fee.",
          "type": "integer",
          "format": "int64"
        },
        "height": {
          "description": "The block height.",
          "type": "integer",
          "format": "int64"
        },
        "difficulty": {
          "description": "The block difficulty.",
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "AccountWalletRequest": {
      "description": "A request representing an account and a wallet.",
      "allOf": [
        {
          "$ref": "#/definitions/AccountId"
        },
        {
          "properties": {
            "wallet": {
              "description": "The wallet name.",
              "type": "string"
            }
          }
        }
      ]
    },
    "AccountWalletPasswordRequest": {
      "description": "A request representing an account and a wallet + password.",
      "allOf": [
        {
          "$ref": "#/definitions/AccountId"
        },
        {
          "$ref": "#/definitions/WalletNamePasswordPair"
        }
      ]
    },
    "AccountStatusViewModel": {
      "description": "A view model representing harvest information.",
      "properties": {
        "status": {
          "type": "string",
          "description": "The status of the account.\n*UNKNOWN - The account status is unknown (needed for ncc).\n*LOCKED - The account is locked.\n*UNLOCKED - The account is unlocked"
        }
      }
    },
...
  },
  "responses": {
    "BadRequest": {
      "description": "Bad request, further information within error response object.",
      "schema": {
        "$ref": "#/definitions/ErrorResponseNcc"
      }
    },
    "Error": {
      "description": "Unexpected error occured.",
      "schema": {
        "$ref": "#/definitions/ErrorResponseHttp"
      }
    }
  }
}